Remove obsolete MySQL dialect
