--- Tests.NHibernate.Spatial.MySQL/MySQLCriteriaFixture.cs
@@ -1,14 +1,0 @@
-using NHibernate.Cfg;
-using NUnit.Framework;
-
-namespace Tests.NHibernate.Spatial
-{
-    [TestFixture]
-    public class MySQLCriteriaFixture : CriteriaFixture
-    {
-        protected override void Configure(Configuration configuration)
-        {
-            TestConfiguration.Configure(configuration);
-        }
-    }
-}
