--- Tests.NHibernate.Spatial.MySQL/MySQLSpatialQueriesFixture.cs
@@ -1,110 +1,0 @@
-using NHibernate;
-using NHibernate.Cfg;
-using NUnit.Framework;
-using Tests.NHibernate.Spatial.RandomGeometries;
-
-namespace Tests.NHibernate.Spatial
-{
-    [TestFixture]
-    public class MySQLSpatialQueriesFixture : SpatialQueriesFixture
-    {
-        protected override void Configure(Configuration configuration)
-        {
-            TestConfiguration.Configure(configuration);
-        }
-
-        protected override string SqlLineStringFilter(string filterString)
-        {
-            return $@"
-SELECT count(*)
-FROM linestringtest
-WHERE MBRIntersects(the_geom, GeomFromText('{filterString}', 4326))
-";
-        }
-
-        protected override string SqlPolygonFilter(string filterString)
-        {
-            return $@"
-SELECT count(*)
-FROM polygontest
-WHERE MBRIntersects(the_geom, GeomFromText('{filterString}', 4326))
-";
-        }
-
-        protected override string SqlMultiLineStringFilter(string filterString)
-        {
-            return $@"
-SELECT count(*)
-FROM multilinestringtest
-WHERE MBRIntersects(the_geom, GeomFromText('{filterString}', 4326))
-";
-        }
-
-        protected override string SqlOvelapsLineString(string filterString)
-        {
-            return $@"
-SELECT count(*)
-FROM linestringtest
-WHERE the_geom IS NOT NULL
-AND Overlaps(PolygonFromText('{filterString}', 4326), the_geom)
-";
-        }
-
-        protected override string SqlIntersectsLineString(string filterString)
-        {
-            return $@"
-SELECT count(*)
-FROM linestringtest
-WHERE the_geom IS NOT NULL
-AND Intersects(PolygonFromText('{filterString}', 4326), the_geom)
-";
-        }
-
-        protected override ISQLQuery SqlIsEmptyLineString(ISession session)
-        {
-            return session.CreateSQLQuery(@"
-SELECT IsEmpty(the_geom) as result
-FROM linestringtest
-WHERE oid = ?
-AND the_geom IS NOT NULL
-")
-                .AddScalar("result", NHibernateUtil.Boolean);
-        }
-
-        protected override ISQLQuery SqlIsSimpleLineString(ISession session)
-        {
-            return session.CreateSQLQuery(@"
-SELECT IsSimple(the_geom) as result
-FROM linestringtest
-WHERE oid = ?
-AND the_geom IS NOT NULL
-")
-                .AddScalar("result", NHibernateUtil.Boolean);
-        }
-
-        protected override ISQLQuery SqlAsBinaryLineString(ISession session)
-        {
-            return session.CreateSQLQuery(@"
-SELECT AsBinary(the_geom) as result
-FROM linestringtest
-WHERE oid = ?
-AND the_geom IS NOT NULL
-")
-                .AddScalar("result", NHibernateUtil.BinaryBlob);
-        }
-
-        #region Unsupported features
-
-        [Test]
-        [Ignore("Provider does not support the Boundary function")]
-        public override void HqlBoundary()
-        { }
-
-        [Test]
-        [Ignore("Provider does not support the Relate function")]
-        public override void HqlRelateLineString()
-        { }
-
-        #endregion
-    }
-}
