--- Tests.NHibernate.Spatial.MySQL/Tests.NHibernate.Spatial.MySQL.csproj
@@ -1,28 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-
-    <RootNamespace>Tests.NHibernate.Spatial</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="nunit" Version="3.13.3" />
-    <PackageReference Include="NUnit3TestAdapter" Version="[3.15.1, 3.16.0)" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="[15.9.2, 16.0)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\NHibernate.Spatial.MySQL\NHibernate.Spatial.MySQL.csproj" />
-    <ProjectReference Include="..\Tests.NHibernate.Spatial\Tests.NHibernate.Spatial.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="appsettings.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
