--- Tests.NHibernate.Spatial.MySQL/appsettings.json
@@ -1,5 +1,0 @@
-{
-  "ConnectionStrings": {
-    "MySQL": "Server=localhost;Port=3306;Database=nhsp_test;User Id=nhsp_test;Password=nhsp_test;"
-  }
-}
